requirements.txt
dj-database-url==0.3.0
Django==1.9.4
djangorestframework==3.3.3
psycopg2==2.6.1
python-dotenv==0.3.0
